Resume Intelligence Analyzer/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.12.5
spacy==3.8.7
pdfplumber==0.11.7
pypdf==5.4.0
python-docx==1.2.0
